+++ .storybook/main.ts
@@ -2,8 +2,16 @@ const config = {
   stories: ['../src/**/*.stories.@(js|jsx|ts|tsx)'],
   addons: ['@storybook/addon-links', '@storybook/addon-docs'],
   framework: {
-    name: '@stencil/storybook-plugin',
+    name: '@storybook/html-vite',
   },
+  managerHead: head => `
+    ${head}
+    <base href="./" />
+  `,
+  previewHead: head => `
+    ${head}
+    <base href="./" />
+  `,
 };
 
 export default config;