+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- [Added handling for icon name](https://github.com/multiversx/mx-sdk-dapp-ui/pull/250)
+- [Updated icons type](https://github.com/multiversx/mx-sdk-dapp-ui/pull/250)
 
 - [Added the Storybook as a live server deployment](https://github.com/multiversx/mx-sdk-dapp-ui/pull/248)