+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Updated icons type](https://github.com/multiversx/mx-sdk-dapp-ui/pull/250)
+
 - [Added the Storybook as a live server deployment](https://github.com/multiversx/mx-sdk-dapp-ui/pull/248)
 
 - [Refactor components in transactions table](https://github.com/multiversx/mx-sdk-dapp-ui/pull/243)